requirements.txt
alpaca-py>=0.21.0
requests>=2.31.0
yfinance>=0.2.36
pandas>=2.2.0
numpy>=1.26.0
schedule>=1.2.0
python-dotenv>=1.0.0
rich>=13.7.0
jinja2>=3.1.0
flask>=3.0.0
